+++ src-tauri/src/entities/mash_steps.rs
@@ -3,7 +3,7 @@
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
 #[sea_orm(table_name = "mash_steps")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false, column_type = "Text")]
@@ -14,11 +14,11 @@ pub struct Model {
     pub name: String,
     #[sea_orm(column_type = "Text")]
     pub r#type: String,
-    pub infuse_amount_l: Option<Decimal>,
-    pub step_temp_c: Decimal,
+    pub infuse_amount_l: Option<f64>,
+    pub step_temp_c: f64,
     pub step_time_min: i32,
     pub ramp_time_min: Option<i32>,
-    pub end_temp_c: Option<Decimal>,
+    pub end_temp_c: Option<f64>,
     pub step_order: i32,
 }
